pyproject.toml
[tool.poetry]
name = "sophie-etl"
version = "0.1.1"
description = "ETL for SOPHIE"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]
[tool.poetry.dependencies]
python = "3.12"
langchain = "0.3.0"
langchain-anthropic = "^0.3.6"
langchain-groq = "0.2.3"
langchain-openai = "^0.3.5"
langchain-deepseek = "^0.1.2"
langchain-ollama = "^0.2.0"
langgraph = "0.2.56"
pandas = "^2.1.0"
numpy = "^1.24.0"
python-dotenv = "1.0.0"
matplotlib = "^3.9.2"
tabulate = "^0.9.0"
colorama = "^0.4.6"
questionary = "^2.1.0"
rich = "^13.9.4"
langchain-google-genai = "^2.0.11"
psycopg2-binary = "^2.9.10"
polygon-api-client = "^1.14.5"
yfinance = "^0.2.35"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
ta = "^0.11.0"
ipykernel = "^6.29.5"
jinja2 = "^3.1.6"
xgboost = "^3.0.1"
backtrader = "^1.9.78.123"
scipy = "^1.15.3"
pandas-market-calendars = "^5.1.0"
shap = "^0.47.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 420
target-version = ['py310']
include = '\.pyi?$'